+++ wordpress/wtyczki/aai-sklep/index.php
@@ -1,0 +1,3 @@
+<?php
+// Cisza. Plik istnieje po to, żeby serwer z włączonym listowaniem
+// katalogów nie pokazywał zawartości wtyczki.